backend/requirements.txt
django >= 4.2.6
python-dotenv >= 1.0.0
psycopg2 >= 2.9.9
djangorestframework >= 3.14.0
# djangorestframework-simplejwt >= 5.3.0

pillow >= 10.1.0
